Rename genome_stats -> genome_attribs
Attributes is a more general name; stats doesn't really fit
what's actually going in the DB
